Corrigin teste da função avg()

+++ operations.py
@@ -34,7 +34,7 @@ class Operations:
         return(self.value1 ** 2)
 
     def avg(self):
-        return(sum(self.value1)/len(self.value1))
+        return float(sum(self.value1) / len(self.value1))
 
     def factorial(self):
         return math.factorial(self.value1)
